requirements.txt
streamlit
pillow
numpy
Groq
pdf2image